# media/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Path-FinderX - API Documentation",
    "version": "1.0.0",
    "description": "API documentation for a path finder application on the town of Bucaramanga, Colombia.",
    "contact": {
      "name": "AngelBLadex",
      "url": "https://portafolio-beta-ebon.vercel.app/",
      "email": "angel.orteg@jala.university"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Node": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 123456789
          },
          "lat": {
            "type": "number",
            "example": 7.123456
          },
          "lon": {
            "type": "number",
            "example": -73.123456
          }
        }
      },
      "RouteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "64b84fa2891cc93a14a57c90"
          },
          "userId": {
            "type": "string",
            "example": "64b84f934f92cd3a14a57c88"
          },
          "name": {
            "type": "string",
            "example": "Ruta segura al colegio"
          },
          "route": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Node"
            }
          },
          "totalDistance": {
            "type": "number",
            "example": 1.25
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-26T14:23:45.000Z"
          }
        }
      },
      "RouteCreateRequest": {
        "type": "object",
        "required": [
          "name",
          "route",
          "totalDistance"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Ruta parque"
          },
          "route": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Node"
            }
          },
          "totalDistance": {
            "type": "number",
            "example": 3.75
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "687b02437e2366cb3eb9bca9"
          },
          "firstName": {
            "type": "string",
            "example": "angel"
          },
          "lastName": {
            "type": "string",
            "example": "ortega"
          },
          "username": {
            "type": "string",
            "example": "angelito_374"
          },
          "email": {
            "type": "string",
            "example": "angelgabrielorteg@gmail.com"
          },
          "role": {
            "type": "string",
            "example": "user"
          },
          "isVerified": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-19T02:26:11.171Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-19T02:26:22.374Z"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Server"
    }
  ],
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "summary": "Create a new user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "firstName",
                  "lastName",
                  "username",
                  "email",
                  "password"
                ],
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Invalid data"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "User login",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "400": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/v1/path/calculate": {
      "post": {
        "summary": "Calculates the optimal route from a starting node to an end node, with optional stops and obstacles",
        "tags": [
          "Path"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "start",
                  "end"
                ],
                "properties": {
                  "start": {
                    "type": "number",
                    "example": 12345
                  },
                  "end": {
                    "type": "number",
                    "example": 67890
                  },
                  "stops": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    },
                    "example": [
                      34567,
                      45678
                    ]
                  },
                  "obstacles": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    },
                    "example": [
                      11111,
                      22222
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Optimal route calculated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "number"
                      },
                      "lat": {
                        "type": "number"
                      },
                      "lon": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid data or route not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/route/user": {
      "get": {
        "summary": "Get paginated routes by current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Route"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "size",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated user routes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RouteResponse"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "currentPage": {
                          "type": "integer",
                          "example": 1
                        },
                        "pageSize": {
                          "type": "integer",
                          "example": 10
                        },
                        "totalCount": {
                          "type": "integer",
                          "example": 25
                        },
                        "totalPages": {
                          "type": "integer",
                          "example": 3
                        },
                        "search": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/route": {
      "post": {
        "summary": "Save a calculated route",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Route"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Route created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/api/v1/route/{id}": {
      "delete": {
        "summary": "Delete a route",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Route"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route deleted"
          },
          "404": {
            "description": "Route not found"
          }
        }
      }
    },
    "/api/v1/user": {
      "get": {
        "summary": "Get paginated list of users with search capability",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "The page number"
          },
          {
            "in": "query",
            "name": "size",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "The number of items per page"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search term (username, first name or last name)"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserResponse"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "currentPage": {
                          "type": "integer",
                          "example": 1
                        },
                        "pageSize": {
                          "type": "integer",
                          "example": 10
                        },
                        "totalCount": {
                          "type": "integer",
                          "example": 1
                        },
                        "totalPages": {
                          "type": "integer",
                          "example": 1
                        },
                        "hasPrevious": {
                          "type": "boolean",
                          "example": false
                        },
                        "hasNext": {
                          "type": "boolean",
                          "example": false
                        },
                        "search": {
                          "type": "string",
                          "example": ""
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/user/{id}": {
      "get": {
        "summary": "Get a user by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User found"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "patch": {
        "summary": "Partially update a user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "Angel"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Ortega"
                  },
                  "username": {
                    "type": "string",
                    "example": "angel_ortega"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "angel@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "MySecurePass123"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated"
          },
          "400": {
            "description": "Invalid data"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "tags": []
}